pyproject.toml
[project]
name = "BDPRIN-projeto"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kellyalvesp", email = "zkellypatriciaalves14@gmail.com"},
]
dependencies = ["django>=5.2.4", "djangorestframework>=3.16.0"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
